rename cmclient => cmcapiclient. scaffold blob and cosmos client code + tests.
